// packages/database/src/pg/models/token_update.ts
export interface PgTokenDistributionUpdate extends commons.BaseSubRecordUpdate {
  totalDeposit?: number | null | Increment;
  totalPaid?: number | null | Increment;
  refundedAmount?: number | null | Increment;
  totalBought?: number | null | Increment;
  reconciled?: boolean | null;
  billPaymentId?: string | null;
  creditPaymentId?: string | null;
  royaltyBillPaymentId?: string | null;
  tokenClaimed?: number | null | Increment;
  lockedForSale?: number | null | Increment;
  sold?: number | null | Increment;
  totalPurchased?: number | null | Increment;
  tokenOwned?: number | null | Increment;
  mintedClaimedOn?: Date | null;
  mintingTransactions?: string[] | null | ArrayUnion<string> | ArrayRemove<string>;
  stakeRewards?: number | null | Increment;
  extraStakeRewards?: number | null | Increment;
  totalUnclaimedAirdrop?: number | null | Increment;
  stakeVoteTransactionId?: string | null;
  stakes_static_amount?: number | null | Increment;
  stakes_static_totalAmount?: number | null | Increment;
  stakes_static_value?: number | null | Increment;
  stakes_static_totalValue?: number | null | Increment;
  stakes_static_stakingMembersCount?: number | null | Increment;
  stakes_dynamic_amount?: number | null | Increment;
  stakes_dynamic_totalAmount?: number | null | Increment;
  stakes_dynamic_value?: number | null | Increment;
  stakes_dynamic_totalValue?: number | null | Increment;
  stakes_dynamic_stakingMembersCount?: number | null | Increment;
  stakeExpiry?: string | any | null;
}